:house: Rename typings file
