+++ templates/tsconfig-test.json.tmpl
@@ -8,5 +8,5 @@
 			"chai": ["../../../types/chai"]
 		}
 	},
-	"include" : ["../../../types/**/*", "./**/*"]
+	"include": ["../../../types/**/*", "./**/*"]
 }
